feat(list-view): support components in item templates
