+++ platform/titanium/runtime/components/list-view/item-template.js
@@ -1,7 +1,3 @@
-import { TitaniumElementRegistry } from 'titanium-vdom';
-
-const registry = TitaniumElementRegistry.getInstance();
-
 export default {
 	name: 'ItemTemplate',
 	props: {
@@ -11,35 +7,11 @@ export default {
 		}
 	},
 
-	render() {
-		this.$parent.addTemplate(this.name, {
-			childTemplates: convertNodesToTemplates(this.$slots.default)
-		});
-		return null;
+	render(h) {
+		return h('detached-view', this.$slots.default);
+	},
+
+	mounted() {
+		this.$parent.registerTemplate(this);
 	}
 };
-
-function convertNodesToTemplates(nodes) {
-	let templates = [];
-	for (let node of nodes) {
-		let meta = registry.getViewMetadata(node.tag);
-		let templateDefinition = {
-			type: meta.typeName
-		};
-		let properties = {};
-		for (let attributeName in node.data.attrs) {
-			const attributeValue = node.data.attrs[attributeName];
-			if (attributeName === 'bindId') {
-				templateDefinition.bindId = attributeValue;
-			} else {
-				properties[attributeName] = attributeValue;
-			}
-		}
-		templateDefinition.properties = properties;
-		if (node.children && node.children.length > 0) {
-			templateDefinition.childTemplates = convertNodesToTemplates(node.children);
-		}
-		templates.push(templateDefinition);
-	}
-	return templates;
-}
