+++ platform/titanium/runtime/components/list-view/list-view.js
@@ -1,3 +1,7 @@
+import { TitaniumElementRegistry } from 'titanium-vdom';
+
+const registry = TitaniumElementRegistry.getInstance();
+
 export default {
 	name: 'ListView',
 
@@ -19,18 +23,24 @@ export default {
 	mounted() {
 		this.$el.setAttribute('templates', this.templates);
 		this.$el.setAttribute('sections', this.sections);
+		const listView = this.$el.titaniumView;
 		const owningView = this.$el.parentElement.titaniumView;
 		if (!owningView) {
 			throw new Error(`Parent ${this.$el.parentElement} of list-view component is not a Titanium element`);
 		}
-		owningView.add(this.$el.titaniumView);
-		this.$el.titaniumView.update
+		owningView.add(listView);
 	},
 
 	methods: {
-		addTemplate(templateName, template) {
-			this.templates[templateName] = template;
+		registerTemplate(itemTemplate) {
+			const templates = [];
+			convertNodesToTemplates(itemTemplate.$el.children, templates);
+			this.templates[itemTemplate.name] = {
+				childTemplates: templates
+			};
+			itemTemplate.$destroy();
 		},
+
 		appendSection(section) {
 			this.sections.push(section);
 			if (this._isMounted) {
@@ -39,3 +49,31 @@ export default {
 		}
 	}
 };
+
+function convertNodesToTemplates(nodes, templates) {
+	if (!nodes) {
+		console.log('No more child nodes to convert, stopping');
+		return;
+	}
+	for (let node of nodes) {
+		console.log(`converting ${node} to template`);
+		let meta = registry.getViewMetadata(node.tagName);
+		let templateDefinition = {
+			type: meta.typeName,
+			childTemplates: []
+		};
+		let properties = {};
+		node.attributes.forEach((attributeValue, attributeName) => {
+			if (attributeName === 'bindId') {
+				templateDefinition.bindId = attributeValue;
+			} else {
+				properties[attributeName] = attributeValue;
+			}
+		});
+		templateDefinition.properties = properties;
+		if (node.children && node.children.length > 0) {
+			convertNodesToTemplates(node.children, templateDefinition.childTemplates);
+		}
+		templates.push(templateDefinition);
+	}
+}
